Relocate LiquidityDialog
