+++ packages/frontend/app/routes/assets.$assetId.add-liquidity.tsx
@@ -1,7 +1,7 @@
 import { type ActionArgs } from '@remix-run/node'
 import { useNavigate } from '@remix-run/react'
 import { v4 } from 'uuid'
-import { LiquidityDialog } from '~/components/ui/LiquidityDialog'
+import { LiquidityDialog } from '~/components/LiquidityDialog'
 import { addAssetLiquidity } from '~/lib/api/asset.server'
 import { messageStorage, setMessageAndRedirect } from '~/lib/message.server'
 import { amountSchema } from '~/lib/validate.server'
